# Conflicts:
#	src/main/java/com/aviumauctores/pioneers/controller/InGameController.java
#	src/main/java/com/aviumauctores/pioneers/service/BuildService.java
#	src/main/resources/com/aviumauctores/pioneers/lang.properties
#	src/main/resources/com/aviumauctores/pioneers/lang_de.properties
#	src/main/resources/com/aviumauctores/pioneers/lang_en.properties
#	src/main/resources/com/aviumauctores/pioneers/views/ingameScreen.fxml

+++ src/main/java/com/aviumauctores/pioneers/controller/GameReadyController.java
@@ -430,7 +430,7 @@ public class GameReadyController extends PlayerListController {
                         String message = errorCodes.get(response.statusCode() + "_game");
                         Platform.runLater(() -> app.showHttpErrorDialog(response.statusCode(), response.error(), message));
                     } else {
-                        app.showErrorDialog(bundle.getString("connection.failed"), bundle.getString("try.again"));
+                        app.showErrorDialog(bundle.getString("smth.went.wrong"), bundle.getString("try.again"));
                     }
                 }));
     }
@@ -448,7 +448,7 @@ public class GameReadyController extends PlayerListController {
                                 String message = errorCodes.get(response.statusCode() + "_member");
                                 Platform.runLater(() -> app.showHttpErrorDialog(response.statusCode(), response.error(), message));
                             } else {
-                                app.showErrorDialog(bundle.getString("connection.failed"), bundle.getString("limit.reached"));
+                                app.showErrorDialog(bundle.getString("smth.went.wrong"), bundle.getString("limit.reached"));
                             }
                         });
 
@@ -477,13 +477,15 @@ public class GameReadyController extends PlayerListController {
             memberAlert.setTitle(bundle.getString("warning"));
             memberAlert.setHeaderText(null);
             Optional<ButtonType> result = memberAlert.showAndWait();
-            if (result.get() == proceedButton) {
-                gameMemberService.deleteMember(userService.getCurrentUserID())
-                        .observeOn(FX_SCHEDULER)
-                        .subscribe();
-            } else {
-                memberAlert.close();
-                return;
+            if (result.isPresent()) {
+                if (result.get() == proceedButton) {
+                    gameMemberService.deleteMember(userService.getCurrentUserID())
+                            .observeOn(FX_SCHEDULER)
+                            .subscribe();
+                } else {
+                    memberAlert.close();
+                    return;
+                }
             }
         }
         gameService.setCurrentGameID(null);
@@ -534,7 +536,7 @@ public class GameReadyController extends PlayerListController {
                                         String message = errorCodes.get(response.statusCode() + "_member");
                                         Platform.runLater(() -> app.showHttpErrorDialog(response.statusCode(), response.error(), message));
                                     } else {
-                                        app.showErrorDialog(bundle.getString("connection.failed"), bundle.getString("limit.reached"));
+                                        app.showErrorDialog(bundle.getString("smth.went.wrong"), bundle.getString("limit.reached"));
                                     }
                                 });
             }
